move some pieces to a separate file
