+++ src/Protocol/DocumentUri.cs
@@ -10,14 +10,27 @@ using static System.IO.Path;
 namespace OmniSharp.Extensions.LanguageServer.Protocol
 {
     /// <summary>
+    /// Uniform Resource Identifier (URI) http://tools.ietf.org/html/rfc3986.
+    /// This class is a simple parser which creates the basic component parts
+    /// (http://tools.ietf.org/html/rfc3986#section-3) with minimal validation
+    /// and encoding.
+    ///
+    /// ```txt
+    ///       foo://example.com:8042/over/there?name=ferret#nose
+    ///       \_/   \______________/\_________/ \_________/ \__/
+    ///        |           |            |            |        |
+    ///     scheme     authority       path        query   fragment
+    ///        |   _____________________|__
+    ///       / \ /                        \
+    ///       urn:example:animal:ferret:nose
+    /// ```
+    /// </summary>
+    /// <summary>
     /// This class describes a document uri as defined by https://microsoft.github.io/language-server-protocol/specifications/specification-current/#uri
     /// </summary>
     /// <remarks>This exists because of some non-standard serialization in vscode around uris and .NET's behavior when deserializing those uris</remarks>
     public partial class DocumentUri : IEquatable<DocumentUri>
     {
-        private static readonly Regex WindowsPath =
-            new Regex(@"^\w(?:\:|%3a)[\\|\/]", RegexOptions.Compiled | RegexOptions.IgnoreCase);
-
         /// <summary>
         /// scheme is the "http' part of 'http://www.msft.com/some/path?query#fragment".
         /// The part before the first colon.
@@ -317,456 +330,18 @@ namespace OmniSharp.Extensions.LanguageServer.Protocol
         /// A default comparer that can be used for equality
         /// </summary>
         public static IEqualityComparer<DocumentUri> Comparer { get; } = new DocumentUriEqualityComparer();
-    }
-
-    partial class DocumentUri
-    {
-        private static readonly Regex SchemePattern = new Regex(@"^\w[[a-zA-Z\d+.-]*$");
-        private  static readonly Regex SingleSlashStart = new Regex(@"^\/");
-        private  static readonly Regex DoubleSlashStart = new Regex(@"^\/\/");
-
-        public static bool IsHighSurrogate(int charCode)
-        {
-            return 0xD800 <= charCode && charCode <= 0xDBFF;
-        }
-
-        public static bool IsLowSurrogate(int charCode)
-        {
-            return 0xDC00 <= charCode && charCode <= 0xDFFF;
-        }
-
-        public static bool IsLowerAsciiHex(int code)
-        {
-            return code >= CharCode.a && code <= CharCode.f;
-        }
-
-        public static bool IsLowerAsciiLetter(int code)
-        {
-            return code >= CharCode.a && code <= CharCode.z;
-        }
-
-        public static bool IsUpperAsciiLetter(int code)
-        {
-            return code >= CharCode.A && code <= CharCode.Z;
-        }
-
-        public static bool IsAsciiLetter(int code)
-        {
-            return IsLowerAsciiLetter(code) || IsUpperAsciiLetter(code);
-        }
-
-        public static void _validateUri(DocumentUri ret, bool? strict)
-        {
-            // scheme, must be set
-            if (string.IsNullOrWhiteSpace(ret.Scheme) && strict == true)
-            {
-                throw new UriFormatException(
-                    $@"Scheme is missing: {{scheme: "", authority: ""{ret.Authority}"", path: ""{ret.Path}"", query: ""${ret.Query}"", fragment: ""{ret.Fragment}""}}");
-            }
-
-            // scheme, https://tools.ietf.org/html/rfc3986#section-3.1
-            // ALPHA *( ALPHA / DIGIT / "+" / "-" / "." )
-            if (!string.IsNullOrWhiteSpace(ret.Scheme) && !SchemePattern.IsMatch(ret.Scheme))
-            {
-                throw new UriFormatException("Scheme contains illegal characters.");
-            }
-
-            // path, http://tools.ietf.org/html/rfc3986#section-3.3
-            // If a URI contains an authority component, then the path component
-            // must either be empty or begin with a slash ("/") character.  If a URI
-            // does not contain an authority component, then the path cannot begin
-            // with two slash characters ("//").
-            if (!string.IsNullOrWhiteSpace(ret.Path))
-            {
-                if (!string.IsNullOrWhiteSpace(ret.Authority))
-                {
-                    if (!SingleSlashStart.IsMatch(ret.Path))
-                    {
-                        throw new UriFormatException(
-                            "If a URI contains an authority component, then the path component must either be empty or begin with a slash (\"/\") character");
-                    }
-                }
-                else
-                {
-                    if (DoubleSlashStart.IsMatch(ret.Path))
-                    {
-                        throw new UriFormatException(
-                            "If a URI does not contain an authority component, then the path cannot begin with two slash characters (\"//\")");
-                    }
-                }
-            }
-        }
 
         // for a while we allowed uris *without* schemes and this is the migration
         // for them, e.g. an uri without scheme and without strict-mode warns and falls
         // back to the file-scheme. that should cause the least carnage and still be a
         // clear warning
-        static string _schemeFix(string scheme, bool? strict)
-        {
-            if (string.IsNullOrWhiteSpace(scheme) && strict != true)
-            {
-                return Uri.UriSchemeFile;
-            }
-
-            return scheme;
-        }
-
-        const string Empty = "";
-        const char Slash = '/';
-        const string StrSlash = "/";
-        static readonly Regex Regexp = new Regex(@"^(([^:/?#]+?):)?(\/\/([^/?#]*))?([^?#]*)(\?([^#]*))?(#(.*))?");
 
         // implements a bit of https://tools.ietf.org/html/rfc3986#section-5
-        static string _referenceResolution(string scheme, string path)
-        {
-            // the slash-character is our "default base' as we don"t
-            // support constructing URIs relative to other URIs. This
-            // also means that we alter and potentially break paths.
-            // see https://tools.ietf.org/html/rfc3986#section-5.1.4
-            if (scheme == Uri.UriSchemeHttps || scheme == Uri.UriSchemeHttp || scheme == Uri.UriSchemeFile)
-            {
-                if (string.IsNullOrWhiteSpace(path))
-                {
-                    path = StrSlash;
-                }
-                else if (path[0] != Slash)
-                {
-                    path = Slash + path;
-                }
-            }
-
-            return path;
-        }
 
 
         // reserved characters: https://tools.ietf.org/html/rfc3986#section-2.2
-        static readonly IDictionary<int, string> EncodeTable = new Dictionary<int, string>() {
-            [CharCode.Colon] = "%3A", // gen-delims
-            [CharCode.Slash] = "%2F",
-            [CharCode.QuestionMark] = "%3F",
-            [CharCode.Hash] = "%23",
-            [CharCode.OpenSquareBracket] = "%5B",
-            [CharCode.CloseSquareBracket] = "%5D",
-            [CharCode.AtSign] = "%40",
-
-            [CharCode.ExclamationMark] = "%21", // sub-delims
-            [CharCode.DollarSign] = "%24",
-            [CharCode.Ampersand] = "%26",
-            [CharCode.SingleQuote] = "%27",
-            [CharCode.OpenParen] = "%28",
-            [CharCode.CloseParen] = "%29",
-            [CharCode.Asterisk] = "%2A",
-            [CharCode.Plus] = "%2B",
-            [CharCode.Comma] = "%2C",
-            [CharCode.Semicolon] = "%3B",
-            [CharCode.Equals] = "%3D",
-
-            [CharCode.Space] = "%20",
-        };
-
-        static string EncodeUriComponentFast(string uriComponent, bool allowSlash)
-        {
-            string res = null;
-            var nativeEncodePos = -1;
-
-            for (var pos = 0; pos < uriComponent.Length; pos++)
-            {
-                var code = uriComponent[pos];
-
-                // unreserved characters: https://tools.ietf.org/html/rfc3986#section-2.3
-                if (
-                    (code >= CharCode.a && code <= CharCode.z)
-                    || (code >= CharCode.A && code <= CharCode.Z)
-                    || (code >= CharCode.Digit0 && code <= CharCode.Digit9)
-                    || code == CharCode.Dash
-                    || code == CharCode.Period
-                    || code == CharCode.Underline
-                    || code == CharCode.Tilde
-                    || (allowSlash && code == CharCode.Slash)
-                    || allowSlash && (pos == 1 || pos == 2) && (
-                        uriComponent.Length >= 3 && uriComponent[0] == CharCode.Slash &&
-                        uriComponent[2] == CharCode.Colon
-                        || uriComponent.Length >= 2 && uriComponent[1] == CharCode.Colon
-                    )
-                )
-                {
-                    // check if we are delaying native encode
-                    if (nativeEncodePos != -1)
-                    {
-                        res += Uri.EscapeDataString(uriComponent.Substring(nativeEncodePos, pos - nativeEncodePos));
-                        nativeEncodePos = -1;
-                    }
-
-                    // check if we write into a new string (by default we try to return the param)
-                    if (res != null)
-                    {
-                        res += uriComponent[pos];
-                    }
-                }
-                else
-                {
-                    // encoding needed, we need to allocate a new string
-                    if (res == null)
-
-                    {
-                        res = uriComponent.Substring(0, pos);
-                    }
-
-                    // check with default table first
-                    if (EncodeTable.TryGetValue(code, out var escaped))
-                    {
-                        // check if we are delaying native encode
-                        if (nativeEncodePos != -1)
-                        {
-                            res += Uri.EscapeDataString(uriComponent.Substring(nativeEncodePos, pos - nativeEncodePos));
-                            nativeEncodePos = -1;
-                        }
-
-                        // append escaped variant to result
-                        res += escaped;
-                    }
-                    else if (nativeEncodePos == -1)
-                    {
-                        // use native encode only when needed
-                        nativeEncodePos = pos;
-                    }
-                }
-            }
-
-            if (nativeEncodePos != -1)
-            {
-                res += Uri.EscapeDataString(uriComponent.Substring(nativeEncodePos));
-            }
-
-            return !string.IsNullOrWhiteSpace(res) ? res : uriComponent;
-        }
-
-        static string EncodeUriComponentMinimal(string path)
-        {
-            string res = null;
-            for (var pos = 0; pos < path.Length; pos++)
-            {
-                var code = path[pos];
-                if (code == CharCode.Hash || code == CharCode.QuestionMark)
-                {
-                    if (res == null)
-                    {
-                        res = path.Substring(0, pos);
-                    }
-
-                    res += EncodeTable[code];
-                }
-                else
-                {
-                    if (res != null)
-                    {
-                        res += path[pos];
-                    }
-                }
-            }
-
-            return res ?? path;
-        }
-
-        /// <summary>
-        /// Compute `fsPath` for the given uri
-        /// </summary>
-        public static string UriToFsPath(DocumentUri uri, bool keepDriveLetterCasing)
-        {
-            string value;
-            if (!string.IsNullOrWhiteSpace(uri.Authority) && uri.Path.Length > 1 && uri.Scheme == "file")
-            {
-                // unc path: file://shares/c$/far/boo
-                value = $"//{uri.Authority}{uri.Path}";
-            }
-            else if (
-                uri.Path.Length >= 3
-                && uri.Path[0] == CharCode.Slash
-                && (uri.Path[1] >= CharCode.A && uri.Path[1] <= CharCode.Z ||
-                    uri.Path[1] >= CharCode.a && uri.Path[1] <= CharCode.z)
-                && uri.Path[2] == CharCode.Colon
-            )
-            {
-                if (!keepDriveLetterCasing)
-                {
-                    // windows drive letter: file:///c:/far/boo
-                    value = char.ToLowerInvariant(uri.Path[1]) + uri.Path.Substring(2);
-                }
-                else
-                {
-                    value = uri.Path.Substring(1);
-                }
-            }
-            else
-            {
-                // other path
-                value = uri.Path;
-            }
-
-            if (WindowsPath.IsMatch(value) || value.StartsWith("//"))
-            {
-                return value.Replace("/", "\\");
-            }
-
-            return value;
-        }
-
-        /// <summary>
-        /// Create the external version of a uri
-        /// </summary>
-        static string _asFormatted(DocumentUri uri, bool skipEncoding)
-        {
-            string Encoder(string p, bool allowSlash)
-            {
-                return !skipEncoding ? EncodeUriComponentFast(p, allowSlash) : EncodeUriComponentMinimal(p);
-            }
-
-            var res = "";
-            var (scheme, authority, path, query, fragment) = uri;
-            if (!string.IsNullOrWhiteSpace(scheme))
-            {
-                res += scheme;
-                res += ":";
-            }
-
-            if (!string.IsNullOrWhiteSpace(authority) || scheme == "file")
-            {
-                res += Slash;
-                res += Slash;
-            }
-
-            if (!string.IsNullOrWhiteSpace(authority))
-            {
-                var idx = authority.IndexOf("@");
-                if (idx != -1)
-                {
-                    // <user>@<auth>
-                    var userinfo = authority.Substring(0, idx);
-                    authority = authority.Substring(idx + 1);
-                    idx = userinfo.IndexOf(":");
-                    if (idx == -1)
-                    {
-                        res += Encoder(userinfo, false);
-                    }
-                    else
-                    {
-                        // <user>:<pass>@<auth>
-                        res += Encoder(userinfo.Substring(0, idx), false);
-                        res += ":";
-                        res += Encoder(userinfo.Substring(idx + 1), false);
-                    }
-
-                    res += "@";
-                }
-
-                authority = authority.ToLowerInvariant();
-                idx = authority.IndexOf(":", StringComparison.Ordinal);
-                if (idx == -1)
-                {
-                    res += Encoder(authority, false);
-                }
-                else
-                {
-                    // <auth>:<port>
-                    res += Encoder(authority.Substring(0, idx), false);
-                    res += authority.Substring(idx);
-                }
-            }
-
-            if (!string.IsNullOrWhiteSpace(path))
-            {
-                // lower-case windows drive letters in /C:/fff or C:/fff
-                if (path.Length >= 3 && path[0] == CharCode.Slash && path[2] == CharCode.Colon)
-                {
-                    var code = path[1];
-                    if (code >= CharCode.A && code <= CharCode.Z)
-                    {
-                        path = $"/{Convert.ToChar(code + 32)}:{path.Substring(3)}"; // "/c:".length == 3
-                    }
-                }
-                else if (path.Length >= 2 && path[1] == CharCode.Colon)
-                {
-                    var code = path[0];
-                    if (code >= CharCode.A && code <= CharCode.Z)
-                    {
-                        path = $"{Convert.ToChar(code + 32)}:{path.Substring(2)}"; // "/c:".length == 3
-                    }
-                }
-
-                // encode the rest of the path
-                res += Encoder(path, true);
-            }
-
-            if (!string.IsNullOrWhiteSpace(query))
-            {
-                res += "?";
-                res += Encoder(query, false);
-            }
-
-            if (!string.IsNullOrWhiteSpace(fragment))
-            {
-                res += "#";
-                res += !skipEncoding ? EncodeUriComponentFast(fragment, false) : fragment;
-            }
-
-            return res;
-        }
 
         // --- decode
-
-        static string DecodeUriComponentGraceful(string str)
-        {
-            try
-            {
-                return Uri.UnescapeDataString(str);
-            }
-            catch
-            {
-                if (str.Length > 3)
-                {
-                    return str.Substring(0, 3) + DecodeUriComponentGraceful(str.Substring(3));
-                }
-                else
-                {
-                    return str;
-                }
-            }
-        }
-
-        static readonly Regex REncodedAsHex =
-            new Regex(@"(%[0-9A-Za-z][0-9A-Za-z])+", RegexOptions.Multiline | RegexOptions.Compiled);
-
-        static string PercentDecode(string str)
-        {
-            if (!REncodedAsHex.IsMatch(str))
-            {
-                return str;
-            }
-
-            return REncodedAsHex.Replace(str, match => DecodeUriComponentGraceful(match.Value));
-        }
-    }
-
-    // var isWindows = RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
-    /// <summary>
-    /// Uniform Resource Identifier (URI) http://tools.ietf.org/html/rfc3986.
-    /// This class is a simple parser which creates the basic component parts
-    /// (http://tools.ietf.org/html/rfc3986#section-3) with minimal validation
-    /// and encoding.
-    ///
-    /// ```txt
-    ///       foo://example.com:8042/over/there?name=ferret#nose
-    ///       \_/   \______________/\_________/ \_________/ \__/
-    ///        |           |            |            |        |
-    ///     scheme     authority       path        query   fragment
-    ///        |   _____________________|__
-    ///       / \ /                        \
-    ///       urn:example:animal:ferret:nose
-    /// ```
-    /// </summary>
-    partial class DocumentUri
-    {
-        static readonly bool IsWindows = RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
 
         /// <summary>
         /// @internal
@@ -774,9 +349,9 @@ namespace OmniSharp.Extensions.LanguageServer.Protocol
         public DocumentUri(string scheme, string authority, string path, string query, string fragment,
             bool? strict = null)
         {
-            Scheme = _schemeFix(scheme, strict);
+            Scheme = SchemeFix(scheme, strict);
             Authority = authority ?? Empty;
-            Path = _referenceResolution(Scheme, path ?? Empty);
+            Path = ReferenceResolution(Scheme, path ?? Empty);
             Query = query ?? Empty;
             Fragment = fragment ?? Empty;
 
@@ -879,14 +454,5 @@ namespace OmniSharp.Extensions.LanguageServer.Protocol
                 components.Fragment ?? string.Empty
             );
         }
-    }
-
-    public struct DocumentUriComponents
-    {
-        public string Scheme { get; set; }
-        public string Authority { get; set; }
-        public string Path { get; set; }
-        public string Query { get; set; }
-        public string Fragment { get; set; }
     }
 }
